msal: Remove python msal bindings
