+++ src/msal/Cargo.toml
@@ -14,7 +14,6 @@ name = "msal"
 path = "src/lib.rs"
 
 [dependencies]
-pyo3 = { version = "^0.18.3", features = ["auto-initialize"] }
 tracing = { workspace = true }
 reqwest = { workspace = true }
 serde = { workspace = true }
